BioMakie API

BioMakie.atomcolors

BioMakie.atomradii

BioMakie.atomradius

BioMakie.atomsizes

BioMakie.backbonebonds

BioMakie.bondshape

BioMakie.bondshapes

BioMakie.covalentbonds

BioMakie.distancebonds

BioMakie.firstlabel

BioMakie.getbonds

BioMakie.getbonds

BioMakie.getinspectorlabel

BioMakie.msavalues

BioMakie.plotmsa

BioMakie.plotmsa

BioMakie.plotmsa

BioMakie.plotmsa!

BioMakie.plotmsa!

BioMakie.plotmsa!

BioMakie.plotstruc

BioMakie.plotstruc!

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.rescolors

BioMakie.sidechainbonds

BioMakie.atomcolors — Method

atomcolors( atoms )

Get a Vector of colors for the atoms. To see all default element and amino acid colorschemes, use getbiocolors(). Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomradii — Method

atomradii( atoms )

Collect atom radii based on element for plotting.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomradius — Method

atomradius( atom )

Collect atom radius based on element for plotting.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( atms )

Get a Vector of sizes for the atoms from a BioStructures.StructuralElementOrList.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.backbonebonds — Method

backbonebonds( chn::BioStructures.Chain ) -> BitMatrix

Returns a matrix of backbone bonds in chn, where Mat[i,j] = 1 if atoms i and j are bonded.

Keyword Arguments:

cutoff –––––- 1.6 # distance cutoff for bonds

BioMakie.bondshape — Method

bondshape( twoatoms )
bondshape( twopoints )

Returns a (mesh) cylinder between two atoms or atomic coordinates.

Keyword Arguments:

bondwidth ––––––- 0.2

BioMakie.bondshapes — Method

bondshapes( structure )
bondshapes( residues )
bondshapes( structure, bondmatrix )
bondshapes( residues, bondmatrix )

Returns a (mesh) cylinder between two atoms or points.

Keyword Arguments:

algo ––––––––– :knowledgebased | :distance, :covalent # unless bondmatrix is given

distance ––––––– 1.9 # unless bondmatrix is given

bondwidth ––––––- 0.2

BioMakie.covalentbonds — Method

covalentbonds( atms ) -> BitMatrix

Returns a matrix of all bonds in atms, where Mat[i,j] = 1 if atoms i and j are bonded.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

extradistance –– 0.14 # fudge factor for better inclusion

H –––––––– true # include bonds with hydrogen atoms

disulfides –––- false # include disulfide bonds

BioMakie.distancebonds — Method

distancebonds( atms ) -> BitMatrix

Returns a matrix of all bonds in atms, where Mat[i,j] = 1 if atoms i and j are bonded.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

hydrogencutoff –- 1.14 # distance cutoff for bonds with hydrogen atoms

H –––––––– true # include bonds with hydrogen atoms

disulfides –––- false # include disulfide bonds

BioMakie.firstlabel — Method

firstlabel( inspectorfunc::Function )
firstlabel( inspectorfunc::Observable{T} ) where {T<:Function}

Show an example of the inspector label function looks like. The position p will not be available to this function, so it will be set to nothing.

BioMakie.getbonds — Method

getbonds( chn::BioStructures.Chain, selectors... ) -> BitMatrix

Returns a matrix of all bonds in chn, where Mat[i,j] = 1 if atoms i and j are bonded.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion

disulfides –––- false # include disulfide bonds

BioMakie.getbonds — Method

getbonds( residues ) -> BitMatrix

Returns a matrix of all bonds in residues, where Mat[i,j] = 1 if atoms i and j are bonded.

Keyword Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion

disulfides –––- false # include disulfide bonds

BioMakie.getinspectorlabel — Method

getinspectorlabel( structure )
getinspectorlabel( residues )
getinspectorlabel( atom )

Get the inspector label function for plotting a 'StructuralElementOrList'.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

BioMakie.msavalues — Function

msavalues( msa::AbstractMatrix, resdict::AbstractDict )::Matrix{Real}

Returns a matrix of numbers according to the given dictionary, where keys are residue letters and values are numbers. This matrix is used as input for plotmsa for the heatmap colors.

Default values for residue letters are from Kidera Factor values. kf 2 is Kidera Factor 2 (size/volume-related). The KF dictionary is in utils.jl.

BioMakie.plotmsa! — Function

plotmsa!( fig, msa, msavalues, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

fig = Figure(resolution = (1100, 400))

plotmsa!( fig::Figure, msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Keyword Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa! — Method

plotmsa!( msamatrix, matrixvals, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

plotmsa!( msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Keyword Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa! — Method

plotmsa!( fig, msamatrix, matrixvals, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

fig = Figure(resolution = (1100, 400))

plotmsa!( fig::Figure, msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Keyword Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( fig, msamatrix, msavalues, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) on a Figure.

BioMakie.plotmsa — Method

plotmsa( msa, msavalues )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

using MIToS.MSA
downloadpfam("PF00062")
msa = MIToS.MSA.read("PF00062.stockholm.gz", Stockholm, 
					generatemapping =true, useidcoordinates=true)
msamatrix, xlabels, ylabels = plottingdata(msa) .|> Observable			
matrixvals = msavalues(msamatrix[]) |> Observable

plotmsa( msa, matrixvals; kwargs... )

Keyword Arguments:

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( msa )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

downloadpfam("PF00062")
msa = MIToS.MSA.read("PF00062.stockholm.gz", Stockholm, 
					generatemapping =true, useidcoordinates=true)

plotmsa( msa; kwargs... )

Keyword Arguments:

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

resdict –––––- kideradict # Dictionary of values (::Dict{String,Float}, "Y" => 1.48) for heatmap.

kf –––––––– 2 # If resdict == kideradict, this is the Kidera Factor. KF2 is size/volume-related.

kwargs... # forwarded to scatter plot

BioMakie.plotstruc! — Method

plotstruc!( fig, structure )

Plot a protein structure(/chain/residues/atoms) into a Figure.

Examples

fig = Figure()

using MIToS.PDB

pdbfile = MIToS.PDB.downloadpdb("2vb1")
struc = MIToS.PDB.read(pdbfile, PDBML) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = @residues struc model "1" chain "A" group "ATOM" residue All
strucplot = plotstruc!(fig, chain_A)

chnatms = @atoms struc model "1" chain "A" group "ATOM" residue All atom All
strucplot = plotstruc!(fig, chnatms)
-------------------------
using BioStructures

struc = retrievepdb("2vb1", dir = "data/") |> Observable
strucplot = plotstruc!(fig, struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = retrievepdb("2hhb", dir = "data/")["A"] |> Observable
strucplot = plotstruc!(fig, chain_A)

Keyword Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors(), or provide a Dict like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plotstruc — Method

plotstruc( structure )
plotstruc( plotdata )

Create and return a Makie Figure for a protein structural element.

Examples

using MIToS.PDB

pdbfile = MIToS.PDB.downloadpdb("2vb1")
struc = MIToS.PDB.read(pdbfile, PDBML) |> Observable
strucplot = plotstruc(struc)

chain_A = @residues struc model "1" chain "A" group "ATOM" residue All
strucplot = plotstruc(chain_A)

chnatms = @atoms struc model "1" chain "A" group "ATOM" residue All atom All
strucplot = plotstruc(chnatms)
-------------------------
using BioStructures

struc = retrievepdb("2vb1", dir = "data/") |> Observable
strucplot = plotstruc(struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
strucplot = plotstruc(struc)

chain_A = retrievepdb("2hhb", dir = "data/")["A"] |> Observable
strucplot = plotstruc(chain_A)

Keyword Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors(), or provide a Dict like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plottingdata — Method

plottingdata( structure )
plottingdata( residues )
plottingdata( atoms )

This function returns an OrderedDict of the main data used for plotting.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Returns:

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => ...)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( msa )

Collects data for plotting (residue string matrix, x labels, and y labels) from a multiple sequence alignment (MSA) object.

The MSA object can be a:

AbstractMultipleSequenceAlignment from MIToS.MSA,

vector of tuples 'Vector{Tuple{String,String}}' from FastaIO,

vector of FASTA records 'Vector{FASTX.FASTA.Record}' from FASTX.

BioMakie.rescolors — Method

rescolors( residues )

Get a Vector of colors for the atoms. To see all default element and amino acid colorschemes, use getbiocolors(). Keyword argument colors takes a Dict which maps residue to color. ("C" => :red)

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

BioMakie.sidechainbonds — Method

sidechainbonds( res::BioStructures.AbstractResidue, selectors... ) -> BitMatrix

Returns a matrix of sidechain bonds in res, where Mat[i,j] = 1 if atoms i and j are bonded.

This function uses 'bestoccupancy' or 'defaultatom' to ensure only one position per atom.

Keyword Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion