API

Index

BioMakie.PDBe_downloader

BioMakie.viewmsa

BioMakie.viewstruc

List

BioMakie.PDBe_downloader — Method

PDBe_downloader(args)

Create and return a Makie Figure for a PDBe data downloader. Source URL: "https://www.ebi.ac.uk/pdbe/api/pdb/entry/{entryname}/{pdbid}

Entries: ["Summary","Molecules","Assoc. Publications", "Ligands","Carbohydrate Polymer","Modified Residues", "Mutated Residues","Secondary Structure","Binding Sites"]

Examples

fig = PDBe_downloader("2vb1")

pdbid = "2vb1" |> Node
fig = PDBe_downloader(pdbid)

Keyword arguments: dir –– Default - "."

BioMakie.viewmsa — Method

viewmsa(msa)

Create and return a Makie Figure for a AbstractMultipleSequenceAlignment.

Examples

using MIToS.MSA
downloadpfam("pf00062")
vm = MIToS.MSA.read("pf00062.stockholm.gz",Stockholm) |> Node
fig1 = viewmsa(vm)

using FastaIO
vm = FastaIO.readfasta("data/fasta1.fas") |> Node
fig1 = viewmsa(vm)

Parameters: sheetsize ––- Dimensions of the msa shown, Default - [40,20] resolution –– Default - (1500, 600) colorscheme –- Default - :viridis positions ––- Residue positions, Default - 1:(length of msa)

BioMakie.viewstruc — Method

viewstruc(args)

Create and return a Makie Figure for a PDB structure.

Examples

using BioStructures
struc = retrievepdb("2vb1", dir = "data\") |> Node
sv = viewstruc(struc)

struc = read("data\2vb1_mutant1.pdb", BioStructures.PDB) |> Node
sv = viewstruc(struc)

Keyword arguments: resolution –– Default - (800,800) atmcolors ––- Default - "element", define your own dict for atoms like: "N" => :blue atmscale ––– Default - 1/3, size adjustment of atom radii figdims –––- Default - [13,10], GridLayout dimensions for structure view: 'fig = Figure(); fig[ 1:(figdims[1]), 1:(figdims[2]) ]'