BioMakie API

BioMakie.atomcolors

BioMakie.atomcolors

BioMakie.atomcolors

BioMakie.atomcolors

BioMakie.atomcolors

BioMakie.atomcolors

BioMakie.atomradii

BioMakie.atomradii

BioMakie.atomradius

BioMakie.atomradius

BioMakie.atomsizes

BioMakie.atomsizes

BioMakie.atomsizes

BioMakie.atomsizes

BioMakie.atomsizes

BioMakie.atomsizes

BioMakie.backbonebonds

BioMakie.covalentbonds

BioMakie.distancebonds

BioMakie.firstlabel

BioMakie.firstlabel

BioMakie.getbonds

BioMakie.getbonds

BioMakie.getbonds

BioMakie.inspectorlabel

BioMakie.inspectorlabel

BioMakie.inspectorlabel

BioMakie.inspectorlabel

BioMakie.inspectorlabel

BioMakie.inspectorlabel

BioMakie.msavalues

BioMakie.plotmsa

BioMakie.plotmsa

BioMakie.plotmsa

BioMakie.plotmsa!

BioMakie.plotmsa!

BioMakie.plotmsa!

BioMakie.plotstruc

BioMakie.plotstruc

BioMakie.plotstruc!

BioMakie.plotstruc!

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.plottingdata

BioMakie.rescolors

BioMakie.rescolors

BioMakie.rescolors

BioMakie.rescolors

BioMakie.sidechainbonds

BioMakie.atomcolors — Method

atomcolors( struc::BioStructures.StructuralElementOrList )

Get a Vector of colors for the atoms in a BioStructures StructuralElementOrList. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomcolors — Method

atomcolors( atms::Vector{MIToS.PDB.PDBAtom} )

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBAtom. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomcolors — Method

atomcolors( resz::Vector{MIToS.PDB.PDBResidue} )

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBResidue. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomcolors — Method

atomcolors( struc::Observable{T} ) where {T<:BioStructures.StructuralElementOrList}

Get a Vector of colors for the atoms in a BioStructures StructuralElementOrList. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomcolors — Method

atomcolors( atms::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBAtom}}

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBAtoms. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomcolors — Method

atomcolors( resz::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBResidue}}

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBResidues. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps element to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors

BioMakie.atomradii — Method

atomradii( atoms::Vector{T} ) where T<:BioStructures.AbstractAtom

Collect atom radii based on element for plotting, from a Vector of BioStructures.AbstractAtoms.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomradii — Method

atomradii( atoms::Vector{T} ) where T<:MIToS.PDB.PDBAtom

Collect atom radii based on element for plotting, from a Vector of MIToS.PDB.PDBAtoms.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomradius — Method

atomradius( atom::BioStructures.Atom )

Collect atom radius based on element for plotting.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomradius — Method

atomradius( atom::MIToS.PDB.PDBAtom )

Collect atom radius based on element for plotting.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( struc::BioStructures.StructuralElementOrList )

Get a Vector of sizes for the atoms from a BioStructures.StructuralElementOrList.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( atms::Vector{MIToS.PDB.PDBAtom} )

Get a Vector of sizes for the atoms from a Vector of MIToS.PDB.PDBAtoms.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( resz::Vector{MIToS.PDB.PDBResidue} )

Get a Vector of sizes for the atoms from a Vector of MIToS.PDB.PDBResidues.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( struc::Observable{T} ) where {T<:BioStructures.StructuralElementOrList}

Get a Vector of sizes for the atoms from a BioStructures.StructuralElementOrList.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( atms::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBAtom}}

Get a Vector of sizes for the atoms from a Vector of MIToS.PDB.PDBAtoms.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.atomsizes — Method

atomsizes( resz::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBResidue}}

Get a Vector of sizes for the atoms from a Vector of MIToS.PDB.PDBResidues.

Keyword Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.backbonebonds — Method

backbonebonds( chn::BioStructures.Chain )

Returns a matrix of backbone bonds in chn.

Optional Arguments:

cutoff –––––- 1.6 # distance cutoff for bonds

BioMakie.covalentbonds — Method

covalentbonds( atms::Vector{T} )

Returns a matrix of all bonds in atms.

Optional Arguments:

extradistance –– 0.14 # fudge factor for better inclusion

H –––––––– true # include bonds with hydrogen atoms

disulfides –––- false # include disulfide bonds

BioMakie.distancebonds — Method

distancebonds( atms::Vector{T} )

Returns a matrix of all bonds in atms.

Optional Arguments:

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

hydrogencutoff –- 1.14 # distance cutoff for bonds with hydrogen atoms

H –––––––– true # include bonds with hydrogen atoms

disulfides –––- false # include disulfide bonds

BioMakie.firstlabel — Method

firstlabel( inspectorfunc::Function )

Show an example of the inspector label function looks like. The position p will not be available to this function, so it will be set to nothing.

BioMakie.firstlabel — Method

firstlabel( inspectorfunc::Observable{T} )

Show an example of the inspector label function looks like. The position p will not be available to this function, so it will be set to nothing.

BioMakie.getbonds — Method

getbonds( chn::BioStructures.Chain )

Returns a matrix of all bonds in atms.

Optional Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion

disulfides –––- false # include disulfide bonds

BioMakie.getbonds — Method

getbonds( resz::Vector{MIToS.PDB.PDBResidue} )

Returns a matrix of all bonds in resz.

Optional Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion

disulfides –––- false # include disulfide bonds

BioMakie.getbonds — Method

getbonds( resz::Vector{BioStructures.AbstractResidue} )

Returns a matrix of all intra-residue bonds in resz. Utilizes sidechainbonds to get bonds within each residue.

Optional Arguments:

kwargs... # passed to sidechainbonds

BioMakie.inspectorlabel — Method

inspectorlabel( struc::BioStructures.StructuralElementOrList )

Get the inspector label function for plotting a BioStructures.StructuralElementOrList. (ProteinStructure, Model, Chain, Residue, Atom, and lists of these)

BioMakie.inspectorlabel — Method

inspectorlabel( atms::Vector{MIToS.PDB.PDBAtom} )

Get the inspector label function for plotting a vector of MIToS.PDB.PDBAtoms.

BioMakie.inspectorlabel — Method

inspectorlabel( resz::Vector{MIToS.PDB.PDBResidue} )

Get the inspector label function for plotting a Vector of MIToS.PDB.PDBResidues.

BioMakie.inspectorlabel — Method

inspectorlabel( struc::Observable{T} ) where {T<:BioStructures.StructuralElementOrList}

Get the inspector label function for plotting a BioStructures StructuralElementOrList. (ProteinStructure, Model, Chain, Residue, Atom, and lists of these)

BioMakie.inspectorlabel — Method

inspectorlabel( atms::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBAtom}}

Get the inspector label function for plotting a vector of MIToS.PDB.PDBAtoms.

BioMakie.inspectorlabel — Method

inspectorlabel( resz::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBResidue}}

Get the inspector label function for plotting a vector of MIToS.PDB.PDBResidues.

BioMakie.msavalues — Function

msavalues( msa::AbstractMatrix, resdict::AbstractDict )::Matrix{Real}

Returns a matrix of numbers according to the given dictionary, where keys are residue letters and values are numbers. This matrix is used as input for plotmsa for the heatmap colors.

Default values for residue letters are from Kidera Factor values. kf 2 is Kidera Factor 2 (size/volume-related). The KF dictionary is in utils.jl.

BioMakie.plotmsa! — Function

plotmsa!( fig, msa, msavalues, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

fig = Figure(resolution = (1100, 400))

plotmsa!( fig::Figure, msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Optional Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa! — Method

plotmsa!( msamatrix, matrixvals, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

plotmsa!( msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Optional Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa! — Method

plotmsa!( fig, msamatrix, matrixvals, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) into a Figure.

Example

fig = Figure(resolution = (1100, 400))

plotmsa!( fig::Figure, msamatrix::Matrix{String}, matrixvals::Matrix{Float32},
			xlabels::Vector{String}, 	
			ylabels::Vector{String};
			kwargs... )

Optional Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

markersize –– 12

colorscheme –- :buda

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( fig, msamatrix, msavalues, xlabels, ylabels )

Plot a multiple sequence alignment (MSA) on a Figure.

BioMakie.plotmsa — Method

plotmsa( msa, msavalues )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

using MIToS.MSA
downloadpfam("PF00062")
msa = MIToS.MSA.read("PF00062.stockholm.gz", Stockholm, 
					generatemapping =true, useidcoordinates=true)
msamatrix, xlabels, ylabels = plottingdata(msa) .|> Observable			
matrixvals = msavalues(msamatrix[]) |> Observable

plotmsa( msa, matrixvals; kwargs... )

Optional Arguments:

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( msa )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

downloadpfam("PF00062")
msa = MIToS.MSA.read("PF00062.stockholm.gz", Stockholm, 
					generatemapping =true, useidcoordinates=true)

plotmsa( msa; kwargs... )

Optional Arguments:

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

resdict –––––- kideradict # Dictionary of values (::Dict{String,Float}, "Y" => 1.48) for heatmap.

kf –––––––– 2 # If resdict == kideradict, this is the Kidera Factor. KF2 is size/volume-related.

kwargs... # forwarded to scatter plot

BioMakie.plotstruc! — Method

plotstruc!( fig, structure )

Plot a protein structure(/chain/residues/atoms) into a Figure.

Examples

fig = Figure()

using MIToS.PDB
pdbfile = MIToS.PDB.downloadpdb("2HHB")
struc = MIToS.PDB.read(pdbfile, PDBML) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = pdb = @residues struc model "1" chain "A" group "ATOM" residue All
strucplot = plotstruc!(fig, chain_A)

chnatms = @atoms res_2hhb model "1" chain "A" group "ATOM" residue All atom All
strucplot = plotstruc!(fig, chnatms)

using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
strucplot = plotstruc!(fig, struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = retrievepdb("2hhb", dir = "data/")["A"] |> Observable
strucplot = plotstruc!(fig, chain_A)

Keyword Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors, or provide a dict for atoms/residues like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plotstruc! — Method

plotstruc!( fig, structure )

Plot a protein structure(/chain/residues/atoms) into a Figure.

Examples

fig = Figure()

using MIToS.PDB
pdbfile = MIToS.PDB.downloadpdb("2HHB")
struc = MIToS.PDB.read(pdbfile, PDBML) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = pdb = @residues struc model "1" chain "A" group "ATOM" residue All
strucplot = plotstruc!(fig, chain_A)

chnatms = @atoms res_2hhb model "1" chain "A" group "ATOM" residue All atom All
strucplot = plotstruc!(fig, chnatms)

using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
strucplot = plotstruc!(fig, struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
strucplot = plotstruc!(fig, struc)

chain_A = retrievepdb("2hhb", dir = "data/")["A"] |> Observable
strucplot = plotstruc!(fig, chain_A)

Keyword Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors, or provide a dict for atoms/residues like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plotstruc — Method

plotstruc( struc::BioStructures.StructuralElementOrList )

Create and return a Makie Figure for a protein structural element.

Examples

using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
sv = plotstruc(struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
sv = plotstruc(struc)

Optional Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors, or provide a dict for atoms/residues like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plotstruc — Method

plotstruc( struc::Observable )

Create and return a Makie Figure for a protein structural element.

Examples

using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
sv = plotstruc(struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
sv = plotstruc(struc)

Optional Arguments:

resolution ––- (800,600)

gridposition –- (1,1) # if an MSA is already plotted, (2,1:3) works well

plottype –––- :covalent, :ballandstick, or :spacefilling

atomcolors ––- elecolors, others in getbiocolors, or provide a dict for atoms/residues like: "N" => :blue

markersize ––- 0.0

markerscale –– 1.0

bondtype –––- :knowledgebased, :covalent, or :distance

distance –––- 1.9 # distance cutoff for covalent bonds

inspectorlabel - :default, or define your own function like: (self, i, p) -> "atom: ... coords: ..."

kwargs... ––– keyword arguments passed to the atom meshscatter

BioMakie.plottingdata — Method

plottingdata( struc::BioStructures.StructuralElementOrList )

Collects data for plotting from a BioStructures.StructuralElementOrList. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => ...)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( msa::AbstractMultipleSequenceAlignment )

Collects data for plotting (residue string matrix, x labels, and y labels).

This function converts an AbstractMultipleSequenceAlignment (from MIToS.MSA) to a matrix of residue characters, x labels, and y labels.

BioMakie.plottingdata — Method

plottingdata( msa::Vector{FASTX.FASTA.Record} )

Collects data for plotting (residue string matrix, x labels, and y labels).

This function converts a Vector{FASTX.FASTA.Record} to a matrix of residue characters, x labels, and y labels.

BioMakie.plottingdata — Method

plottingdata( atms::Vector{MIToS.PDB.PDBAtom} )

Collects data for plotting from a Vector of MIToS.PDB.PDBAtoms. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => nothing)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( resz::Vector{MIToS.PDB.PDBResidue} )

Collects data for plotting from a Vector of MIToS.PDB.PDBResidues. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => ...)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( msa::Vector{Tuple{String,String}} )

Collects data for plotting (residue string matrix, x labels, and y labels).

This function converts a Vector{Tuple{String,String}} (from FastaIO) to a matrix of residue characters, x labels, and y labels.

BioMakie.plottingdata — Method

plottingdata( struc::Observable{T} ) where {T<:BioStructures.StructuralElementOrList}

Collects data for plotting from a BioStructures.StructuralElementOrList. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => ...)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( atms::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBAtom}}

Collects data for plotting from a Vector of MIToS.PDB.PDBAtoms. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => nothing)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.plottingdata — Method

plottingdata( resz::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBResidue}}

Collects data for plotting from a Vector of MIToS.PDB.PDBResidues. This function returns an OrderedDict of Observables that are the main data used for plotting, for ease of use and consistency.

OrderedDict("atoms" => ..., 
            "coords" => ..., 
            "colors" => ...,
            "sizes" => ...,
            "bonds" => ...)

Keyword Arguments:

colors –––- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals, :bas, :ballandstick

BioMakie.rescolors — Method

rescolors( struc::BioStructures.StructuralElementOrList )

Get a Vector of colors for the atoms from a BioStructures.StructuralElementOrList. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps residue to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

BioMakie.rescolors — Method

rescolors( resz::Vector{MIToS.PDB.PDBResidue} )

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBResidues. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps residue to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

BioMakie.rescolors — Method

rescolors( struc::Observable{T} ) where {T<:BioStructures.StructuralElementOrList}

Get a Vector of colors for the atoms from a BioStructures.StructuralElementOrList. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps residue to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

BioMakie.rescolors — Method

rescolors( resz::Observable{T} ) where {T<:Vector{MIToS.PDB.PDBResidue}}

Get a Vector of colors for the atoms from a Vector of MIToS.PDB.PDBResidues. To see all default element and amino acid colorschemes, use the getbiocolors() function. Keyword argument colors takes a Dict which maps residue to color. ("C" => :red)

Keyword Arguments:

colors –- elecolors | Options - elecolors, aquacolors, shapelycolors, maecolors

BioMakie.sidechainbonds — Method

sidechainbonds( res::BioStructures.AbstractResidue, selectors... )

Returns a matrix of sidechain bonds in res.

Optional Arguments:

algo ––––––- :knowledgebased # (:distance, :covalent) algorithm to find bonds

H –––––––– true # include bonds with hydrogen atoms

cutoff –––––- 1.9 # distance cutoff for bonds between heavy atoms

extradistance –– 0.14 # fudge factor for better inclusion