BioMakie API

BioMakie.atomcoords

BioMakie.atomradii

BioMakie.getplottingdata

BioMakie.msavalues

BioMakie.plotmsa

BioMakie.plotmsa

BioMakie.plotmsa!

BioMakie.plotstruc

BioMakie.plotstruc!

BioMakie.atomcoords — Method

atomcoords(atoms)

Convenience function for collecting atom coordinates for plotting.

BioMakie.atomradii — Method

atomradii(atoms)

Collect atom radii for plotting. Uses BioStructures to get radii based on atomic element.

Optional Arguments:

radiustype –- :covalent | Options - :cov, :covalent, :vdw, :vanderwaals

BioMakie.getplottingdata — Method

getplottingdata( msa )::Tuple{Matrix{String}, Vector{String}, Vector{String}}

Collects data for plotting (residue string matrix, x labels, and y labels).

This function converts an AbstractMultipleSequenceAlignment (from MIToS.MSA), or a Vector{Tuple{String,String}} (from FastaIO), or a Vector{FASTX.FASTA.Record} to a matrix of residue characters, x labels, and y labels.

BioMakie.msavalues — Function

msavalues( msa::AbstractMatrix, resdict::AbstractDict )::Matrix{Real}

Returns a matrix of numbers according to the given dictionary, where keys are residue letters and values are numbers. This matrix is used as input for plotmsa for the heatmap colors.

Default values for residue letters are from Kidera Factor values. kf 2 is Kidera Factor 2 (size/volume-related). The KF dictionary is in utils.jl.

BioMakie.plotmsa! — Method

plotmsa!( fig, msa, msavalues )

Plot a multiple sequence alignment (MSA) into a Figure.

Examples

fig = Figure(resolution = (1100, 400))

plotmsa!( fig::Figure, msamatrix::Matrix{String}, matrixvals::Matrix{Float32};
		xlabels = xlabels1, 	
		ylabels = ylabels1,
		kwargs... )

Optional Arguments:

xlabels –––- {1:height}

ylabels –––- {1:width}

sheetsize ––- [40,20]

gridposition – (1,1)

colorscheme –- :viridis

markersize –– (11.0,11.0)

markercolor –- :black

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( msa, msavalues )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

plotmsa( msamatrix::Matrix{String}, 
		 matrixvals::Matrix{Float32};
		 xlabels = xlabel::Vector{String}, 	
		 ylabels = ylabel::Vector{String}, 
		 kwargs... )

Optional Arguments:

xlabels –––––- {1:height}

ylabels –––––- {1:width}

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

returnobservables - true # Return Observables for interaction.

kwargs... # forwarded to scatter plot

BioMakie.plotmsa — Method

plotmsa( msa )

Plot a multiple sequence alignment (MSA). Returns a Figure, or a Figure and Observables for interaction.

Examples

plotmsa( msamatrix::Matrix{String};
         xlabels = xlabel::Vector{String}, 	
         ylabels = ylabel::Vector{String}, 
         kwargs... )

Optional Arguments:

xlabels –––––- {1:height}

ylabels –––––- {1:width}

resolution –––– (1100, 400)

sheetsize ––––- [40,20]

gridposition ––– (1,1)

colorscheme –––- :viridis

resdict –––––- kideradict # Dictionary of values (::Dict{String,Float}, "Y" => 1.48) for heatmap.

kf –––––––– 2 # If resdict == kideradict, this is the Kidera Factor. KF2 is size/volume-related.

returnobservables - true # Return Observables for interaction.

kwargs... # forwarded to scatter plot

BioMakie.plotstruc! — Method

plotstruc!( fig, structure )

Plot a protein structure into a Figure. Position can be specified with the gridposition keyword argument.

Examples

fig = Figure()
using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
sv = plotstruc!(fig, struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
sv = plotstruc!(fig, struc)

Optional Arguments:

selectors ––- [standardselector]

resolution –– (800,800)

gridposition – (1,1)

plottype ––– :ballandstick, another option is :spacefilling

atomcolors –– elecolors, another option is aquacolors, or define your own dict for atoms like: "N" => :blue

BioMakie.plotstruc — Method

plotstruc(structure)

Create and return a Makie Figure for a protein structure, along with the structure wrapped in an Observable if it wasn't Observable already.

Examples

using BioStructures
struc = retrievepdb("2vb1", dir = "data/") |> Observable
sv = plotstruc(struc)

struc = read("data/2vb1_mutant1.pdb", BioStructures.PDB) |> Observable
sv = plotstruc(struc)

Optional Arguments:

selectors ––- [standardselector]

resolution –– (800,800)

gridposition – (1,1)

plottype ––– :ballandstick, another option is :spacefilling

atomcolors –– elecolors, another option is aquacolors, or define your own dict for atoms like: "N" => :blue